+++ contracts/InterestRateModel.sol
@@ -16,7 +16,11 @@ contract InterestRateModel {
      * @param borrows The total borrows of the asset in the market
      * @return Success or failure and the supply interest rate per block scaled by 10e18
      */
-    function getSupplyRate(address asset, uint cash, uint borrows) public view returns (uint, uint);
+    function getSupplyRate(
+        address asset,
+        uint256 cash,
+        uint256 borrows
+    ) public view returns (uint256, uint256);
 
     /**
      * @notice Gets the current borrow interest rate based on the given asset, total cash and total borrows
@@ -27,5 +31,9 @@ contract InterestRateModel {
      * @param borrows The total borrows of the asset in the market
      * @return Success or failure and the borrow interest rate per block scaled by 10e18
      */
-    function getBorrowRate(address asset, uint cash, uint borrows) public view returns (uint, uint);
+    function getBorrowRate(
+        address asset,
+        uint256 cash,
+        uint256 borrows
+    ) public view returns (uint256, uint256);
 }
