+++ test/AlkemiEarnVerified/AlkemiEarnVerifiedTest_Borrow.sol
@@ -7,15 +7,17 @@ import "./AlkemiEarnVerifiedTest.sol";
  * @dev This tests the money market with tests for borrow.
  */
 contract AlkemiEarnVerifiedTest_Borrow is AlkemiEarnVerifiedTest {
-
     function testBorrow_MarketSupported() public {
         address token = address(this); // must be this
         initializer();
         addKYCAdmin(msg.sender);
         addCustomerKYC(msg.sender);
 
-        uint err = borrow(token, 10);
-        Assert.equal(uint(Error.MARKET_NOT_SUPPORTED), err, "should have returned Error.MARKET_NOT_SUPPORTED");
+        uint256 err = borrow(token, 10);
+        Assert.equal(
+            uint256(Error.MARKET_NOT_SUPPORTED),
+            err,
+            "should have returned Error.MARKET_NOT_SUPPORTED"
+        );
     }
-
 }
