+++ test/AlkemiEarnVerified/AlkemiEarnVerifiedTest_CalculateAmountSeize2.sol
@@ -6,15 +6,20 @@ import "./AlkemiEarnVerifiedWithPriceTest.sol";
 /*
  * @dev This tests the money market with tests for calculateAmountSeize
  */
-contract AlkemiEarnVerifiedTest_CalculateAmountSeize2 is AlkemiEarnVerifiedWithPriceTest {
-
+contract AlkemiEarnVerifiedTest_CalculateAmountSeize2 is
+    AlkemiEarnVerifiedWithPriceTest
+{
     function testCalculateAmountSeize_OverflowOnMaxUintBorrowPrice() public {
         liquidationDiscount = Exp({mantissa: 10**16}); // 10% discount
 
         Exp memory underwaterAssetPrice = Exp({mantissa: 2**256 - 1});
         Exp memory collateralPrice = Exp({mantissa: mantissaOne * 550});
 
-        (Error err, uint result) = calculateAmountSeize(underwaterAssetPrice, collateralPrice, 5 * 10**18);
+        (Error err, uint256 result) = calculateAmountSeize(
+            underwaterAssetPrice,
+            collateralPrice,
+            5 * 10**18
+        );
         Assert.equal(0, result, "default value");
 
         assertError(Error.INTEGER_OVERFLOW, err, "overflows");
@@ -26,7 +31,11 @@ contract AlkemiEarnVerifiedTest_CalculateAmountSeize2 is AlkemiEarnVerifiedWithP
         Exp memory underwaterAssetPrice = Exp({mantissa: mantissaOne});
         Exp memory collateralPrice = Exp({mantissa: mantissaOne});
 
-        (Error err, uint result) = calculateAmountSeize(underwaterAssetPrice, collateralPrice, 2**256 - 1);
+        (Error err, uint256 result) = calculateAmountSeize(
+            underwaterAssetPrice,
+            collateralPrice,
+            2**256 - 1
+        );
         Assert.equal(0, result, "default value");
 
         assertError(Error.INTEGER_OVERFLOW, err, "overflows");
