+++ test/AlkemiEarnVerified/AlkemiEarnVerifiedTest_Oracle2.sol
@@ -7,7 +7,6 @@ import "./AlkemiEarnVerifiedWithPriceTest.sol";
  * @dev This tests the money market with tests for oracle activities, part 2
  */
 contract AlkemiEarnVerifiedTest_Oracle2 is AlkemiEarnVerifiedWithPriceTest {
-
     // Here we are letting the AlkemiEarnVerified be its own oracle.
     function testFetchAssetPrice_Oracle() public {
         oracle = address(this);
@@ -18,10 +17,10 @@ contract AlkemiEarnVerifiedTest_Oracle2 is AlkemiEarnVerifiedWithPriceTest {
         assetPrices[asset2].mantissa = 10;
 
         (Error err1, Exp memory price1) = fetchAssetPrice(asset1);
-        Assert.equal(0, uint(err1), "err1");
+        Assert.equal(0, uint256(err1), "err1");
 
         (Error err2, Exp memory price2) = fetchAssetPrice(asset2);
-        Assert.equal(0, uint(err2), "err2");
+        Assert.equal(0, uint256(err2), "err2");
 
         Assert.equal(5, price1.mantissa, "price1 mantissa");
         Assert.equal(10, price2.mantissa, "price2 mantissa");
